Adds a TabsResponsive Component
Breaks out the logic that converts from tabs to collapsibles into
a TabsResponsive component. Moves the shared code between Tabs and
TabsResponsive to a TabsLogic mixin used by both.
