+++ example/index.html
@@ -47,7 +47,7 @@
     }
 
     React.render(
-      <Tabs onTabActive={logActiveTab} collapsibleSpeed={1200}>
+      <TabsResponsive onTabActive={logActiveTab} collapsibleSpeed={1200}>
         <Tab title="Frogs" id="funkyFrogs" onTabActive={secretFrogLog}>
           <p>Frogs are a diverse and largely carnivorous group of short-bodied, tailless amphibians composing the order Anura (Ancient Greek an-, without + oura, tail). The oldest fossil "proto-frog" appeared in the early Triassic of Madagascar, but molecular clock dating suggests their origins may extend further back to the Permian, 265 million years ago. Frogs are widely distributed, ranging from the tropics to subarctic regions, but the greatest concentration of species diversity is found in tropical rainforests. There are approximately 4,800 recorded species, accounting for over 85% of extant amphibian species. They are also one of the five most diverse vertebrate orders.</p>
 
@@ -77,12 +77,12 @@
 
           <p>In literature and legend, the salamander is associated with fire, being supposedly unharmed by the flames, while clothes made from its skins or 'wool' were believed to be incombustible. More plausibly, salamanders were said to be intensely poisonous. Despite this, salamander brandy, a drink prepared by dunking live salamanders in fermenting fruit juices, is reputed to have hallucinogenic and aphrodisiac properties. The salamander's ability to regenerate lost body parts is being investigated and research is ongoing into any applications this may have for human medicine.</p>
         </Tab>
-      </Tabs>,
+      </TabsResponsive>,
       document.getElementById('tabCollapsibleExample')
     );
 
     React.render(
-      <Tabs convertToCollapsibleGroup={false}>
+      <Tabs>
         <Tab title="Frogs">
           <p>Frogs are a diverse and largely carnivorous group of short-bodied, tailless amphibians composing the order Anura (Ancient Greek an-, without + oura, tail). The oldest fossil "proto-frog" appeared in the early Triassic of Madagascar, but molecular clock dating suggests their origins may extend further back to the Permian, 265 million years ago. Frogs are widely distributed, ranging from the tropics to subarctic regions, but the greatest concentration of species diversity is found in tropical rainforests. There are approximately 4,800 recorded species, accounting for over 85% of extant amphibian species. They are also one of the five most diverse vertebrate orders.</p>
