+++ src/index.js
@@ -1,5 +1,6 @@
 require('./scss/styles.scss');
 
 module.exports.Tabs = require('./components/Tabs');
+module.exports.TabsResponsive = require('./components/TabsResponsive');
 module.exports.Tab = require('./components/Tab');
 module.exports.TabDivider = require('./components/TabDivider');
